topaz-backend-main/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
Authlib==1.6.4
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
deprecation==2.1.0
distro==1.9.0
fastapi==0.116.2
groq==0.31.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
packaging==25.0
postgrest==2.19.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.19.0
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
storage3==2.19.0
StrEnum==0.4.15
stripe==12.5.1
supabase==2.19.0
supabase-auth==2.19.0
supabase-functions==2.19.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
